Landing Page
I have designed The landing page using HTML and CSS the page layout having #header, top navigation bar, body and footer. the body designed with padding and layers of padding aligned the text and images using CSS. The Top Navigation Bar is having button and used hover to change the color of the button when activated. The footer is having helpdesk containing the mail id to complaint any issues of the web page from the users.